프로젝트합체

+++ Assets/01.Scripts/JeonSiseong/Skills/ISkillDamageable.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface ISkillDamageable
 {
-    void TakeSkillDamage(int damage);
+    void TakeSkillDamage(float damage);
 }
